OTP_Auth/otp_auth/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
passlib==1.7.4
PyJWT==1.7.1
pyotp==2.3.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
